I have fixed the function ret value

+++ stdlib/include/io.h
@@ -6,8 +6,8 @@ struct str_view_t {
   uint64_t size;
 };
 typedef struct str_view_t str_view;
-int selfputchar(unsigned char c);
-int selfputcharerr(unsigned char c);
-int selfprint(str_view s);
-int printmulti(unsigned char c, int64_t n);
-int geterrno();
+uint64_t selfputchar(unsigned char c);
+uint64_t selfputcharerr(unsigned char c);
+uint64_t selfprint(str_view s);
+uint64_t printmulti(unsigned char c, int64_t n);
+uint64_t geterrno();
